Add NotificationContext and improv Navbar
