+++ src/components/dashboard/UserDashboard.js
@@ -1,5 +1,4 @@
-import { Link } from 'react-router-dom';
-import { useState } from 'react';
+import UserNav from '../pages/user/UserNav';
 
 
 const UserDashboard = () => {
@@ -8,16 +7,7 @@ const UserDashboard = () => {
     <section>
       <div className="menu-header">
         <h2>User Dashboard</h2>
-        <nav>
-          <ul>
-            <li><Link to="/user-dashboard">Home</Link></li>
-            <li><Link to="/user-profile">Profile</Link></li>
-            <li><Link to="/user-events">Events</Link></li>
-            <li><Link to="/user-requests">Requests</Link></li>
-            <li><Link to="/user-notifications">Notifications</Link></li>
-            <li><Link to="/logout">Logout</Link></li>
-          </ul>
-        </nav>
+        <UserNav />
       </div>
     </section>
   );
